Creation/Destruction of Users Fully Implemented
IN THEORY - HAVEN'T TESTED THIS SHIT YET

The program maintains a MAP of users. For LOGIN feature it checks if the password matches the given user id and returns true if it does. It also stores the username into a static data member and that acts as the 'current logged in user'
This static username can henceforth be used to interact with machines (machine interaction not yet implemented).
If the password is incorrect, the function returns false;
If there is no user of that name, then the function returns false;
In both false cases a statement is printed in cout.

Create new user checks if the user exists already exists, and prints an error line if it already exists and returns false
If not, it allocates new memory for map, creates a new user based on the information given, and returns true;
(should i change the current instance user to the one just created?)

Delete new user first checks for password and returns false if the password doesn't match after printing an error message.
If the password matches, then it erases the given user from the map, prints a statement, and returns true.